subc_test reduce

+++ subc.cpp
@@ -2,14 +2,14 @@
 
 using namespace intx;
 
-[[gnu::noinline]] static auto sub(const uint256& x, const uint256& y)
+[[gnu::noinline]] static auto sub(const uint128& x, const uint128& y)
 {
     return x - y;
 }
 
 int main()
 {
-    if (sub(0, 1) != ~uint256{})
+    if (sub(0, 1) != ~uint128{})
         return 1;
 
     return 0;
